corrected bug in line 584
If idx_handler.stream_both is set then, the l2cap_request_can_send_now_event shall be called for the first device when the my_a2dp_handle_timeout is called. Because in the a2dp_demo_send_media_packet function the l2cap_request_can_send_now_event is called for the second device. Currently it is hard coded. But it works.